requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
clickhouse-driver==0.2.6
python-multipart==0.0.6
python-jose==3.3.0
pandas==2.1.3
pydantic==2.5.2
python-dotenv==1.0.0 